+++ lib/transform.js
@@ -1,206 +1,15 @@
-const resolveConfig = require('tailwindcss/resolveConfig');
-const { CSSStyleDeclaration } = require('../patched-lib/CSSStyleDeclaration.js');
-const { createRounder, getBreakPoints, createTouplesConverter } = require('./utils');
 const { parseColor, parseCss } = require('./parsers');
+const normalizer = require('./normalizer');
+const { isVariable, omitIf } = require('./utils');
 const allProperties = require('cssstyle/lib/allProperties');
 const isMatchWith = require('lodash.ismatchwith');
-const flow = require('lodash.flow');
 const euclideanDistance = require('euclidean-distance');
+const memoize = require('memoize-one');
+
+const colorProps = Array.from(allProperties).filter((prop) => prop.includes('color'));
+const colorPropsSet = new Set(colorProps);
 
 async function transform(inputCss, tailwindCss, options) {
-    const resolvedConfig = resolveConfig(options.TAILWIND_CONFIG);
-
-    const normalizeFontSize = createTouplesConverter({
-        props: ['font-size'],
-        convertValue: createRounder({
-            breakpoints: getBreakPoints(resolvedConfig.theme.fontSize, options),
-            options,
-        }),
-    });
-
-    const normalizeLineHeight = createTouplesConverter({
-        props: ['line-height'],
-        convertValue: createRounder({
-            breakpoints: getBreakPoints(resolvedConfig.theme.lineHeight, options),
-            options,
-        }),
-    });
-
-    const normalizeLetterSpacing = createTouplesConverter({
-        props: ['letter-spacing'],
-        convertValue: createRounder({
-            breakpoints: getBreakPoints(resolvedConfig.theme.letterSpacing, options),
-            options,
-        }),
-    });
-
-    const normalizeBorderRadius = createTouplesConverter({
-        props: ['border-radius'],
-        convertValue: createRounder({
-            breakpoints: getBreakPoints(resolvedConfig.theme.borderRadius, options).filter((num) => num < 100),
-            bailFn: (num) => {
-                // this must be a full round value
-                if (num > 100) {
-                    return options.FULL_ROUND;
-                }
-            },
-            options,
-        }),
-    });
-
-    const colorProps = Array.from(allProperties).filter((prop) => prop.includes('color'));
-    const colorPropsSet = new Set(colorProps);
-
-    const normalizeColorValues = createTouplesConverter({
-        props: colorProps,
-        convertValue: (value) => {
-            const rgba = parseColor(value);
-            if (rgba !== null) {
-                return `rgba(${rgba.join(', ')})`;
-            }
-
-            return value;
-        },
-    });
-
-    const normalizeBorderColorProperties = createTouplesConverter({
-        props: ['border-top-color', 'border-right-color', 'border-bottom-color', 'border-left-color'],
-        convertProp: () => 'border-color',
-    });
-
-    const normalizeBorderStyleProperties = createTouplesConverter({
-        props: ['border-top-style', 'border-right-style', 'border-bottom-style', 'border-left-style'],
-        convertProp: () => 'border-style',
-    });
-
-    const normalizeWidth = createTouplesConverter({
-        props: ['width'],
-        convertValue: createRounder({
-            breakpoints: getBreakPoints(resolvedConfig.theme.width, options),
-            options,
-        }),
-    });
-    const normalizeHeight = createTouplesConverter({
-        props: ['height'],
-        convertValue: createRounder({
-            breakpoints: getBreakPoints(resolvedConfig.theme.height, options),
-            options,
-        }),
-    });
-    const normalizeMargin = createTouplesConverter({
-        props: ['margin', 'margin-top', 'margin-right', 'margin-bottom', 'margin-left'],
-        convertValue: createRounder({
-            breakpoints: getBreakPoints(resolvedConfig.theme.margin, options),
-            options,
-        }),
-    });
-    const normalizePadding = createTouplesConverter({
-        props: ['padding', 'padding-top', 'padding-right', 'padding-bottom', 'padding-left'],
-        convertValue: createRounder({
-            breakpoints: getBreakPoints(resolvedConfig.theme.padding, options),
-            options,
-        }),
-    });
-    const normalizeGap = createTouplesConverter({
-        props: ['gap'],
-        convertValue: createRounder({
-            breakpoints: getBreakPoints(resolvedConfig.theme.gap, options),
-            options,
-        }),
-    });
-
-    const normalizeBorderWidth = createTouplesConverter({
-        props: ['border-width', 'border-top-width', 'border-right-width', 'border-bottom-width', 'border-left-width'],
-        convertValue: createRounder({
-            breakpoints: getBreakPoints(resolvedConfig.theme.borderWidth, options),
-            options,
-        }),
-    });
-
-    const normalizeCssMap = flow([
-        normalizeLineHeight,
-        normalizeLetterSpacing,
-        normalizeFontSize,
-
-        normalizeColorValues,
-
-        normalizeBorderRadius,
-        normalizeBorderWidth,
-        normalizeBorderColorProperties,
-        normalizeBorderStyleProperties,
-
-        normalizeWidth,
-        normalizeHeight,
-        normalizeMargin,
-        normalizePadding,
-        normalizeGap,
-    ]);
-
-    // ///////
-    // ///////
-    // ///////
-    // ///////
-    // ///////
-    // ///////
-    // ///////
-    // ///////
-    // ///////
-    // ///////
-
-    function omitIf(obj, ...fns) {
-        return Object.fromEntries(Object.entries(obj).filter((touple) => !fns.some((fn) => fn(touple))));
-    }
-
-    function isVariable([prop]) {
-        return prop.startsWith('--');
-    }
-
-    function normalizeDictOfTouples(dict, fn) {
-        return Object.fromEntries(
-            Object.entries(dict).map(([twClass, touples]) => {
-                return [twClass, fn(touples)];
-            }),
-        );
-    }
-
-    function resolveLocalVariables(touples) {
-        const variables = touples.filter(isVariable);
-
-        return touples.map(([prop, value]) => {
-            const resolvedValue = variables.reduce((str, [varN, varV]) => str.split(`var(${varN})`).join(varV), value);
-            return [prop, resolvedValue];
-        });
-    }
-
-    function normalizeShorthands(touples) {
-        const declaration = new CSSStyleDeclaration();
-
-        touples.forEach(([prop, value]) => {
-            declaration.setProperty(prop, value);
-        });
-
-        return Object.entries(declaration.getNonShorthandValues());
-    }
-
-    async function createNormalizedClasses(css) {
-        const variants = await parseCss(css);
-
-        return Object.fromEntries(
-            Object.entries(variants).map(([variant, classesJson]) => {
-                const resolvedLocalVariables = normalizeDictOfTouples(classesJson, resolveLocalVariables);
-                const normalizedShorthands = normalizeDictOfTouples(resolvedLocalVariables, normalizeShorthands);
-                const normalizedCssValues = normalizeDictOfTouples(normalizedShorthands, normalizeCssMap);
-                return [variant, normalizeDictOfTouples(normalizedCssValues, Object.fromEntries)];
-            }),
-        );
-    }
-
-    // ////////
-    // ////////
-    // ////////
-    // ////////
-    // ////////
-    // ////////
 
     function isSubset(parent, child, strict) {
         const a = omitIf(parent, isVariable);
@@ -284,8 +93,8 @@ async function transform(inputCss, tailwindCss, options) {
         });
     }
 
-    const tailwindNormalized = await createNormalizedClasses(tailwindCss);
-    const inputNormalized = await createNormalizedClasses(inputCss);
+    const tailwindNormalized = normalizer(await parseCss(tailwindCss), options);
+    const inputNormalized = normalizer(await parseCss(inputCss), options);
 
     const BAZ = Object.entries(inputNormalized)
         .flatMap(([variant, inputNormalized]) => {
