+++ lib/utils.js
@@ -52,6 +52,16 @@ const createTouplesConverter = ({ props, convertProp = (x) => x, convertValue = 
         });
 };
 
+function omitIf(obj, ...fns) {
+    return Object.fromEntries(Object.entries(obj).filter((touple) => !fns.some((fn) => fn(touple))));
+}
+
+function isVariable([prop]) {
+    return prop.startsWith('--');
+}
+
 module.exports.getBreakPoints = getBreakPoints;
 module.exports.createRounder = createRounder;
 module.exports.createTouplesConverter = createTouplesConverter;
+module.exports.omitIf = omitIf;
+module.exports.isVariable = isVariable;
