+++ package.json
@@ -12,6 +12,7 @@
     "lodash.isequal": "^4.5.0",
     "lodash.ismatchwith": "^4.4.0",
     "lodash.merge": "^4.6.2",
+    "memoize-one": "^5.1.1",
     "parse-unit": "^1.0.1",
     "postcss": "^7.0.32",
     "postcss-safe-parser": "^4.0.2",
